+++ src/Application/emc.camus.application/Common/SortDirection.cs
@@ -1,0 +1,13 @@
+namespace emc.camus.application.Common;
+
+/// <summary>
+/// Specifies the direction of sorting.
+/// </summary>
+public enum SortDirection
+{
+    /// <summary>Ascending order.</summary>
+    Asc,
+
+    /// <summary>Descending order.</summary>
+    Desc
+}