+++ src/Test/emc.camus.api.test/Controllers/AuthControllerTests.cs
@@ -160,6 +160,7 @@ public class AuthControllerTests
             .Setup(s => s.GetGeneratedTokensAsync(
                 It.IsAny<PaginationParams>(),
                 It.IsAny<GeneratedTokenFilter>(),
+                It.IsAny<GeneratedTokenSortParams>(),
                 It.IsAny<CancellationToken>()))
             .ReturnsAsync(pagedResult);
 
@@ -187,6 +188,7 @@ public class AuthControllerTests
             .Setup(s => s.GetGeneratedTokensAsync(
                 It.IsAny<PaginationParams>(),
                 It.IsAny<GeneratedTokenFilter>(),
+                It.IsAny<GeneratedTokenSortParams>(),
                 It.IsAny<CancellationToken>()))
             .ReturnsAsync(pagedResult);